requirements.txt
fastapi==0.95.1
uvicorn==0.23.0
transformers==4.30.0
torch==2.1.0
Pillow==9.5.0
ultralytics
pillow
torch>=1.10.0
ultralytics>=8.0.0
numpy<2.0.0
opencv-python>=4.5.5
fastapi>=0.95.0
uvicorn>=0.21.0
click>=8.0.4
typing-extensions>=4.0.0
pydantic>=1.10.0
py7zr
dill
gunicorn==20.1.0
python-multipart